+++ README.md
@@ -28,12 +28,14 @@ npm run build:dev
 ```
 
 ## Структура проекта
-- `src/api` — определения типов публичного API;
-- `src/hooks` — бизнес-логика компонента: управление прямоугольниками, рендеринг, фон и контекстное меню;
-- `src/rendering` — низкоуровневые функции отрисовки фона, сегментов и лупы;
-- `src/utils` — утилиты для геометрии, преобразования координат и работы с canvas;
-- `src/pkg` — React-компонент `ImageHighlighterModule`, собирающий все части вместе;
-- `src/index.ts` — точка входа библиотеки.
+- `src/shared` — переиспользуемые утилиты и определения типов, независимые от конкретных фич;
+- `src/features/highlighter` — доменный модуль размечателя изображений с компонентами, хуками, моделями и рендерами:
+  - `components` — React-компоненты публичного уровня;
+  - `hooks` — бизнес-логика и контроллер состояния;
+  - `model` — функции работы с данными (сортировка, фабрики сегментов);
+  - `renderers` — низкоуровневые процедуры отрисовки.
+- `src/pkg` — совместимость с прежним путём импорта, переэкспортирующая основной компонент;
+- `src/index.ts` — публичная точка входа в библиотеку.
 
 ## Публикация пакета
 ```bash
