+++ src/features/highlighter/hooks/useBackgroundCanvas.ts
@@ -1,7 +1,7 @@
 import { useEffect, useRef } from 'react'
-import { Size } from 'api/types'
-import { createOffscreenCanvas } from 'utils/canvas'
-import { drawBackground } from 'rendering/background'
+import { Size } from 'shared/api/types'
+import { createOffscreenCanvas } from 'shared/canvas'
+import { drawBackground } from 'features/highlighter/renderers/backgroundRenderer'
 
 export const useBackgroundCanvas = (size: Size, image: HTMLImageElement | null) => {
         const offscreenRef = useRef<HTMLCanvasElement | null>(null)
