+++ src/features/highlighter/index.ts
@@ -1,0 +1,3 @@
+export { ImageHighlighterModule } from './components/ImageHighlighter'
+export { defaultSortOptions, stableSort } from './model/sorting'
+export type { SortAxis, SortDirection, SortOptions } from 'shared/api/types'
