+++ src/features/highlighter/renderers/rectanglesRenderer.ts
@@ -1,5 +1,5 @@
-import { HandlePosition, Rect, Size } from 'api/types'
-import { rectToPixels } from 'utils/geometry'
+import { HandlePosition, Rect, Size } from 'shared/api/types'
+import { rectToPixels } from 'shared/geometry'
 
 export interface DrawRectanglesOptions {
         ctx: CanvasRenderingContext2D
