+++ src/index.ts
@@ -1,3 +1,4 @@
-export { ImageHighlighterModule } from './pkg'
-export type { ImageHighlighterRef } from 'api/types'
-export type { ExtendedImageHighlighterProps, ImageHighlighterProps } from 'api/types'
+export { ImageHighlighterModule } from './features/highlighter'
+export type { ImageHighlighterRef } from 'shared/api/types'
+export type { ExtendedImageHighlighterProps, ImageHighlighterProps } from 'shared/api/types'
+export type { SortAxis, SortDirection, SortOptions } from './features/highlighter'
